backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
PyJWT==2.8.0
python-multipart==0.0.6
tenacity==8.2.3
prometheus-client==0.19.0
kubernetes==28.1.0
python-dotenv==1.0.0
structlog==23.3.0
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0
black==23.12.1
ruff==0.1.11
